@@ -16,18 +16,16 @@
         }
     ],
     "require": {
-        "php": "^8.0|^8.1|^8.2",
-        "illuminate/contracts": "^9.0|^10.0",
-        "laravel/dusk": "^7.0"
-    },
-    "conflict": {
-        "laravel/framework": "<9.15.0"
+        "php": "^8.1|^8.2",
+        "illuminate/contracts": "^10.0",
+        "laravel/dusk": "^7.0",
+        "spatie/invade": "^1.1"
     },
     "require-dev": {
         "laravel/pint": "^1.0",
         "nesbot/carbon": "^2.66",
         "nunomaduro/collision": "^6.0",
-        "orchestra/testbench": "^7.0|^8.0",
+        "orchestra/testbench": "^8.0",
         "pestphp/pest": "^1.21",
         "pestphp/pest-plugin-laravel": "^1.1",
         "phpunit/phpunit": "^9.5|^10.0"
